:recycle: 컴포넌트 이름과 구조 변경 및 Menu 컴포넌트 분리
