+++ src/components/Menu/styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const NavBar = styled.ul`
+  & li {
+    margin-bottom: 1rem;
+  }
+`;
